Move definition of unit registry to validation.py
Prep for deleting the utils.py file
